requirements.txt
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
Cython==3.0.3
gitdb==4.0.10
github==1.2.7
GitPython==3.1.37
idna==3.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
numpy==1.26.0
pandas==2.1.1
pyproject-toml==0.0.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
toml==0.10.2
tzdata==2023.3
urllib3==2.0.6
wheel==0.41.2
